some fixes and warning disables in ui/ for building on win x64
- Some |int|s that should be |ptrdiff_t|s.
- Icon structures are packing to known size (same signs).
- GetWindowLong needs to be GetWindowLongPtr to work on 32 & 64.

R=ben@chromium.org, jschuh@chromium.org
BUG=167187, 166496

Review URL: https://chromiumcodereview.appspot.com/11889017

git-svn-id: svn://svn.chromium.org/chrome/trunk/src@176860 0039d316-1c4b-4281-b951-d872f2087c98

Former-commit-id: 1a7f42769b898d79643c817cd741d872377fc364

+++ ui/base/dragdrop/os_exchange_data_provider_win.cc
@@ -964,10 +964,8 @@ static STGMEDIUM* GetStorageForFileDescriptor(
   FILEGROUPDESCRIPTOR* descriptor = locked_mem.get();
   descriptor->cItems = 1;
   descriptor->fgd[0].dwFlags = FD_LINKUI;
-  wcsncpy_s(descriptor->fgd[0].cFileName,
-            MAX_PATH,
-            file_name.c_str(),
-            std::min(file_name.size(), MAX_PATH - 1u));
+  wcsncpy_s(descriptor->fgd[0].cFileName, MAX_PATH, file_name.c_str(),
+            std::min(file_name.size(), static_cast<size_t>(MAX_PATH - 1u)));
 
   STGMEDIUM* storage = new STGMEDIUM;
   storage->tymed = TYMED_HGLOBAL;
